+++ packages/riverpod_lint_flutter_test/test/goldens/stateless_ref.dart
@@ -23,3 +23,6 @@ int incorrectlyTyped(
 ) {
   return 0;
 }
+
+@riverpod
+external int scoped();